% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC content of the DNA sequences across a sliding window}
\arguments{
\item{sequences_df}{A data frame with the column named \code{sequences} that contains DNA sequences as strings.}

\item{window}{An integer specifying the the size of sliding window for GC content calculation}
}
\value{
A data frame with additional column named \code{gc_content_windows} that contains the GC content values
for each window in each sequence. The column is a list where each element contains the GC content for each
window of the corresponding sequence.
}
\description{
This function calculates the GC content of DNA sequences across sliding windows of a specified size.
The GC content is computed for each window, and the results are stored in a new column in the input data frame.
}
\examples{
# Example Data Frame with DNA Sequences
sequences_df <- data.frame(
  sequences = c(
    "AGCTGCGCGTATCGTACGCGATCGTATCGCGATCGTATCGCG",
    "GGCGCGCTAGCTCGAGTCGCGCGGCTCGATAGCTCGTACGTAG"
  ),
  stringsAsFactors = FALSE  # Ensures strings are not converted to factors
)

# Calculate GC content with a sliding window size of 10
result <- gc_content(sequences_df, window = 10)

# Print the result
print(result)

}
