% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA files}
\usage{
read_fasta(path = NA)
}
\arguments{
\item{path}{Path to the FASTA file with DNA sequences}
}
\value{
A data frame with sequences
}
\description{
Read FASTA files
}
